refactor(HomeNavigation): fixing export for Completed Challenges
