+++ android/app/build.gradle
@@ -140,7 +140,7 @@ android {
         applicationId "com.yomamobile"
         minSdkVersion rootProject.ext.minSdkVersion
         targetSdkVersion rootProject.ext.targetSdkVersion
-        versionCode 33
+        versionCode 34
         versionName "0.0.1"
         multiDexEnabled true
     }
