+++ src/modules/CompletedChallenges/index.ts
@@ -1,3 +1,5 @@
+import CompletedChallenges from './CompletedChallenges'
 import CompletedChallengesWidget from './Widget/CompletedChallengesWidget.container'
 
+export default CompletedChallenges
 export { CompletedChallengesWidget }
